+++ Sources/SpyBug/Views/ReportProblemImagePicker.swift
@@ -8,8 +8,6 @@
 import SwiftUI
 import SnapPix
 
-import SwiftUI
-
 struct ReportProblemImagePicker: View {
     @Binding var problemUIImages: [UIImage]
     
@@ -25,17 +23,4 @@ struct ReportProblemImagePicker: View {
     
 }
 
-struct ImagePickerLabel: View {
-    var body: some View {
-        RoundedRectangle(cornerRadius: 20)
-            .fill(.white.opacity(0.3))
-            .frame(width: 100, height: 100)
-            .shadow(color: Color(.shadow), radius: 8, x: 4, y: 4)
-            .overlay(
-                Image(systemName: "plus")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 30, height: 30)
-            )
-    }
-}
+
