--- Sources/SpyBug/Views/RoundedPicture.swift
@@ -1,21 +1,0 @@
-//
-//  SwiftUIView.swift
-//  SpyBug
-//
-//  Created by Szymon Wnuk on 25/11/2024.
-//
-
-import SwiftUI
-
- struct RoundedPicture: View {
-    
-//    var reportType: ReportType = .bug
-     var body: some View {
-        Text("")
-        
-    }
-}
-
-#Preview {
-    RoundedPicture()
-}
